requirements.txt
streamlit
pymongo
streamlit-aggrid
st-annotated-text